move DeriveKey() into hdwalletderive.go, remove unnecessary code
