+++ ioctl/cmd/hdwallet/hdwalletderive.go
@@ -7,14 +7,20 @@
 package hdwallet
 
 import (
+	"bytes"
 	"fmt"
+	"io/ioutil"
 
+	ecrypt "github.com/ethereum/go-ethereum/crypto"
+	hdwallet "github.com/miguelmota/go-ethereum-hdwallet"
 	"github.com/spf13/cobra"
 
+	"github.com/iotexproject/go-pkgs/crypto"
 	"github.com/iotexproject/iotex-address/address"
 	"github.com/iotexproject/iotex-core/ioctl/config"
 	"github.com/iotexproject/iotex-core/ioctl/output"
 	"github.com/iotexproject/iotex-core/ioctl/util"
+	"github.com/iotexproject/iotex-core/pkg/util/fileutil"
 )
 
 // Multi-language support
@@ -24,8 +30,8 @@ var (
 		config.Chinese: "通过派生key生成钱包账号",
 	}
 	hdwalletDeriveCmdUses = map[config.Language]string{
-		config.English: "derive id1/id2[/id3]",
-		config.Chinese: "derive id1/id2[/id3]",
+		config.English: "derive id1/id2/id3",
+		config.Chinese: "derive id1/id2/id3",
 	}
 )
 
@@ -54,16 +60,68 @@ func hdwalletDerive(path string) error {
 		return output.NewError(output.InputError, "failed to get password", err)
 	}
 
-	_, pri, err := DeriveKey(account, change, index, password)
+	addr, _, err := DeriveKey(account, change, index, password)
 	if err != nil {
 		return err
 	}
-
-	addr, err := address.FromBytes(pri.PublicKey().Hash())
-	if err != nil {
-		return output.NewError(output.ConvertError, "failed to convert public key into address", err)
-	}
-	output.PrintResult(fmt.Sprintf("address: %s\n", addr.String()))
+	output.PrintResult(fmt.Sprintf("address: %s\n", addr))
 
 	return nil
+}
+
+// DeriveKey derives the key according to path
+func DeriveKey(account, change, index uint32, password string) (string, crypto.PrivateKey, error) {
+	// derive key as "m/44'/304'/account'/change/index"
+	hdWalletConfigFile := config.ReadConfig.Wallet + "/hdwallet"
+	if !fileutil.FileExists(hdWalletConfigFile) {
+		return "", nil, output.NewError(output.InputError, "Run 'ioctl hdwallet create' to create your HDWallet first.", nil)
+	}
+
+	enctxt, err := ioutil.ReadFile(hdWalletConfigFile)
+	if err != nil {
+		return "", nil, output.NewError(output.InputError, "failed to read config", err)
+	}
+
+	enckey := util.HashSHA256([]byte(password))
+	dectxt, err := util.Decrypt(enctxt, enckey)
+	if err != nil {
+		return "", nil, output.NewError(output.InputError, "failed to decrypt", err)
+	}
+
+	dectxtLen := len(dectxt)
+	if dectxtLen <= 32 {
+		return "", nil, output.NewError(output.ValidationError, "incorrect data", nil)
+	}
+
+	mnemonic, hash := dectxt[:dectxtLen-32], dectxt[dectxtLen-32:]
+	if !bytes.Equal(hash, util.HashSHA256(mnemonic)) {
+		return "", nil, output.NewError(output.ValidationError, "password error", nil)
+	}
+
+	wallet, err := hdwallet.NewFromMnemonic(string(mnemonic))
+	if err != nil {
+		return "", nil, err
+	}
+
+	derivationPath := fmt.Sprintf("%s/%d'/%d/%d", DefaultRootDerivationPath, account, change, index)
+	path := hdwallet.MustParseDerivationPath(derivationPath)
+	walletAccount, err := wallet.Derive(path, false)
+	if err != nil {
+		return "", nil, output.NewError(output.InputError, "failed to get account by derive path", err)
+	}
+
+	private, err := wallet.PrivateKey(walletAccount)
+	if err != nil {
+		return "", nil, output.NewError(output.InputError, "failed to get private key", err)
+	}
+	prvKey, err := crypto.BytesToPrivateKey(ecrypt.FromECDSA(private))
+	if err != nil {
+		return "", nil, output.NewError(output.InputError, "failed to Bytes private key", err)
+	}
+
+	addr, err := address.FromBytes(prvKey.PublicKey().Hash())
+	if err != nil {
+		return "", nil, output.NewError(output.ConvertError, "failed to convert public key into address", err)
+	}
+	return addr.String(), prvKey, nil
 }
